fix: settings tab switching and migration sync issues
- Move SETTINGS_TAB enum to shared tabs.ts (server component can't
  import from 'use client' module)
- Remove incorrect 'use server' directive from page.tsx (it's a page
  component, not a Server Action)
- Disable SequelizeAdapter auto-sync to prevent duplicate column errors
  during migrations
