+++ src/app/settings/tabs.ts
@@ -1,0 +1,7 @@
+export enum SETTINGS_TAB {
+	INSPECTION = 'INSPECTION',
+	CYLINDERS = 'CYLINDERS',
+	NOTIFICATIONS = 'NOTIFICATIONS',
+	USERS = 'USERS',
+	AUDIT_LOG = 'AUDIT_LOG',
+}
